feature/ESE remove spring boot 2.0 test change package names in test classes
