+++ pom.xml
@@ -84,7 +84,7 @@
         <!-- Dependency Versions -->
         <commons-io.version>2.8.0</commons-io.version>
         <!--when updating elasticsearch versions, also update "ElasticsearchContainer" version in EmbeddedElasticsearchExtension-->
-        <elasticsearch.version>7.6.2</elasticsearch.version>
+        <elasticsearch.version>7.5.0</elasticsearch.version>
         <reflections.version>0.9.12</reflections.version>
         <testcontainers.elasticsearch.version>1.15.0</testcontainers.elasticsearch.version>
     </properties>
