--- tests/test-spring-boot-2.0/src/main/resources/es/migration/V001.00__createTemplateWithIndexMapping.http
@@ -1,35 +1,0 @@
-PUT _template/test_1
-Content-Type: application/json
-
-{
-  "index_patterns" : [
-    "test_*"
-  ],
-  "order" : 1,
-  "version" : 1,
-  "settings" : {
-    "number_of_shards" : 1
-  },
-  "mappings" : {
-    "dynamic" : "strict",
-    "properties" : {
-      "doc" : {
-        "dynamic" : false,
-        "properties" : {}
-      },
-      "searchable" : {
-        "dynamic" : false,
-        "properties" : {
-          "version" : {
-            "type" : "keyword",
-            "ignore_above" : 20,
-            "similarity" : "boolean"
-          },
-          "locked" : {
-            "type" : "boolean"
-          }
-        }
-      }
-    }
-  }
-}
