--- tests/test-spring-boot-2.0/src/main/resources/es/migration/V001.01__addDocument.http
@@ -1,17 +1,0 @@
-PUT /test_1/_doc/1?refresh
-Content-Type: application/json
-
-{
-  "searchable": {
-    "version": "1",
-    "locked": false
-  },
-  "doc": {
-    "version": "1",
-    "locked": false,
-    "success": true,
-    "a": "a a a",
-    "b": true,
-    "c": "c"
-  }
-}
