+++ tests/test-spring-boot-2.1-scriptsInJarFile/pom.xml
@@ -17,7 +17,7 @@
         <java.version>1.8</java.version>
 
         <commons-io.version>2.8.0</commons-io.version>
-        <elasticsearch.version>7.6.2</elasticsearch.version>
+        <elasticsearch.version>7.5.0</elasticsearch.version>
         <testcontainers.elasticsearch.version>1.15.0</testcontainers.elasticsearch.version>
     </properties>
