--- tests/test-spring-boot-2.1-scriptsInJarFile/src/main/java/com/senacor/elasticsearch/evolution/springboot21/Application.java
@@ -1,13 +1,0 @@
-package com.senacor.elasticsearch.evolution.springboot21;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-
-}
