+++ tests/test-spring-boot-2.1-scriptsInJarFile/src/test/java/com/senacor/elasticsearch/evolution/springboot23/ApplicationTests.java
@@ -1,4 +1,4 @@
-package com.senacor.elasticsearch.evolution.springboot20;
+package com.senacor.elasticsearch.evolution.springboot23;
 
 import org.apache.http.HttpHost;
 import org.elasticsearch.client.RestClient;
@@ -22,7 +22,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 @SpringBootTest(properties = {"spring.elasticsearch.rest.uris=http://localhost:" + ApplicationTests.ELASTICSEARCH_PORT})
 public class ApplicationTests {
 
-    static final int ELASTICSEARCH_PORT = 18753;
+    static final int ELASTICSEARCH_PORT = 18769;
 
     @Autowired
     private EsUtils esUtils;
@@ -39,8 +39,7 @@ public class ApplicationTests {
     @TestConfiguration
     static class Config {
         @Bean(destroyMethod = "stop")
-        public ElasticsearchContainer elasticsearchContainer(@Value("${elasticsearch.version:7.6.2}") String esVersion) {
-
+        public ElasticsearchContainer elasticsearchContainer(@Value("${elasticsearch.version:7.5.0}") String esVersion) {
             ElasticsearchContainer container = new ElasticsearchContainer(DockerImageName
                     .parse("docker.elastic.co/elasticsearch/elasticsearch-oss")
                     .withTag(esVersion)
