+++ tests/test-spring-boot-2.2/src/main/java/com/senacor/elasticsearch/evolution/springboot22/Application.java
@@ -1,4 +1,4 @@
-package com.senacor.elasticsearch.evolution.springboot20;
+package com.senacor.elasticsearch.evolution.springboot22;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
