+++ tests/test-spring-boot-2.3/src/main/java/com/senacor/elasticsearch/evolution/springboot23/Application.java
@@ -1,4 +1,4 @@
-package com.senacor.elasticsearch.evolution.springboot20;
+package com.senacor.elasticsearch.evolution.springboot23;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
