+++ tests/test-spring-boot-2.3/src/test/java/com/senacor/elasticsearch/evolution/springboot23/EsUtils.java
@@ -1,4 +1,4 @@
-package com.senacor.elasticsearch.evolution.springboot20;
+package com.senacor.elasticsearch.evolution.springboot23;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
