+++ tests/test-spring-boot-2.4/src/main/resources/es/migration/V001.00__createTemplateWithIndexMapping.http
@@ -2,33 +2,31 @@ PUT _template/test_1
 Content-Type: application/json
 
 {
-  "index_patterns": [
+  "index_patterns" : [
     "test_*"
   ],
-  "order": 1,
-  "version": 1,
-  "settings": {
-    "number_of_shards": 1
+  "order" : 1,
+  "version" : 1,
+  "settings" : {
+    "number_of_shards" : 1
   },
-  "mappings": {
-    "_doc": {
-      "dynamic": "strict",
-      "properties": {
-        "doc": {
-          "dynamic": false,
-          "properties": {}
-        },
-        "searchable": {
-          "dynamic": false,
-          "properties": {
-            "version": {
-              "type": "keyword",
-              "ignore_above": 20,
-              "similarity": "boolean"
-            },
-            "locked": {
-              "type": "boolean"
-            }
+  "mappings" : {
+    "dynamic" : "strict",
+    "properties" : {
+      "doc" : {
+        "dynamic" : false,
+        "properties" : {}
+      },
+      "searchable" : {
+        "dynamic" : false,
+        "properties" : {
+          "version" : {
+            "type" : "keyword",
+            "ignore_above" : 20,
+            "similarity" : "boolean"
+          },
+          "locked" : {
+            "type" : "boolean"
           }
         }
       }
