+++ tests/test-spring-boot-2.4/src/test/java/com/senacor/elasticsearch/evolution/springboot24/EsUtils.java
@@ -1,4 +1,4 @@
-package com.senacor.elasticsearch.evolution.springboot21;
+package com.senacor.elasticsearch.evolution.springboot24;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
